Descaler Fix and Improvements, scaling deadline, many new API endpoints (#21)
* Descaler Fix and Improvements, scaling deadline, many new API endpoints (jobs, services)

* clippy check fix

* add rust-toolchain

+++ scheduler/src/background/service_descaler.rs
@@ -35,7 +35,7 @@ pub async fn service_descaler_handler(
                 }
             };
 
-            if let Err(e) = scaler.scale_down(&service, u64::MAX).await {
+            if let Err(e) = scaler.scale_down(&service, i32::MAX).await {
                 error!("ServiceScalerError: {:?}", e);
                 continue;
             };
